[Movements] Implement all aerial and ground-based movements (#99)
* Define all movement behaviors

* Implement all movement behaviors

* Add unit tests for movement behaviors

* Address comments

* Fix style of local constant variable

* Inherit from TestBase instead of AgentTestBase in movement tests

+++ Assets/Scripts/Interceptors/Interceptor.cs
@@ -49,7 +49,7 @@ public class Interceptor : Agent {
 
     // Calculate the boost acceleration.
     float boostAcceleration =
-        (float)((staticConfig.BoostConfig?.BoostAcceleration ?? 0) * Constants.kGravity);
+        (staticConfig.BoostConfig?.BoostAcceleration ?? 0) * Constants.kGravity;
     Vector3 boostAccelerationVector = boostAcceleration * transform.forward;
 
     // Add the PN acceleration to the boost acceleration.
